Add unit-tests for value type related helpers
